Update Bluetooth.md
If you are pairing with a device that requires manual entry of pin, such as with an Arduino Bluetooth Shield, then in the Device Manager conetext you will be requested to enter the pin which you can do. Also, where a pin is not required you can use the Default IoT Core app to pair.