Fixed MD typo in link to Device Management Client Repo
The link to the Device Management Client Repo had an MD (markdown) typo. Specifically, the link should have the form [visible-description](URL). The left-paren '(' was incorrectly a right paren ')', causing the link to fail to render correctly.  The fix is to simply replace the incorrect right-paren with the correct left-paren.