Update windows-iot-enterprise.md
Edits for general readability and conformity to MS style and corporate standards. Please keep in mind that, for localization reasons, links to pages that do not require a locale code to render must be coded as locale-agnostic (that is, they must not contain "/en-us" in the URL). Be sure to scrub for that throughout. And let me know if you have questions!